[課題１０]MyBatisを使ってCRUDのRead処理を実装 ## 実装内容 - students_coursesの全件検索を表示しました
## 動作確認
### 1. 全件検索成功時
[スクリーンショット]

## 工夫した点・学んだこと
- 開始日と終了日がNULLになってしまっていたが、原因がキャメルケースであったがことが分かった。
